공지 파일 등록기능 완료

+++ src/main/java/com/kh/moida/notice/Notice.java
@@ -16,7 +16,7 @@ public class Notice {
     private String noticeContent;
     private Date noticeDate;
     private int noticeCount;
-    private int fileId;
+    private Long  fileId;
 
     private String fileConvert;
